Fix examples/sc2autosave.py my moving sc2reader.scripts.utils.Formatter
