+++ sc2reader/readers.py
@@ -1857,7 +1857,7 @@ class GameEventsReader_34784(GameEventsReader_27950):
                 25: (
                     None,
                     self.command_manager_reset_event,
-                ),  # Re-using this old number
+                ),  # Reusing this old number
                 61: (None, self.trigger_hotkey_pressed_event),
                 103: (CommandManagerStateEvent, self.command_manager_state_event),
                 104: (
